NP-001: added CRUD methods to dev-branch

+++ src/main/java/org/newsportal/dao/entity/Article.java
@@ -10,8 +10,8 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class Article {
-    private int id;
+    private long id;
     private String title;
     private String content;
-    private int userId;
+    private long userId;
 }
